Address PR #7 review comments for e2e helper extraction
